+++ src/hooks/init/api.js
@@ -27,7 +27,7 @@ import { generateIcon } from "../../lib/icons.js";
 import DDBProxy from "../../lib/DDBProxy.js";
 import DDBCompanion from "../../parser/companions/DDBCompanion.js";
 import DDBCompanionFactory from "../../parser/companions/DDBCompanionFactory.js";
-import { configureCustomAAForCondition, addSaveAdvantageToTarget } from "../../effects/helpers.js";
+import { configureCustomAAForCondition, addSaveAdvantageToTarget, findContainedTokensInTemplate } from "../../effects/helpers.js";
 import SETTINGS from "../../settings.js";
 
 function resetSecrets() {
@@ -110,8 +110,9 @@ export function registerApi() {
     DDBCompanion,
     DDBCompanionFactory,
     effects: {
-      configureCustomAAForCondition,
       addSaveAdvantageToTarget,
+      configureCustomAAForCondition,
+      findContainedTokensInTemplate,
     },
     debug: {
       start: debugStart,